Function stage2::prepare_heap

pub(crate) fn prepare_heap(
    kernel_region: MemoryRegion<PhysAddr>,
    loaded_kernel_pregion: MemoryRegion<PhysAddr>,
    loaded_kernel_vregion: MemoryRegion<VirtAddr>,
    platform: &dyn SvsmPlatform,
    config: &SvsmConfig<'_>,
) -> Result<(MemoryRegion<VirtAddr>, MemoryRegion<PhysAddr>), SvsmError>

Maps any remaining memory between the end of the kernel image and the end of the allocated kernel memory region as heap space. Exclude any memory reserved by the configuration.

Panics

Panics if the allocated kernel region (kernel_region) is not sufficient to host the loaded kernel region (loaded_kernel_pregion) plus memory reserved for configuration.