Crate stage2

Modules

boot_stage2

Structs

KernelPageTablePage

Stage2BootLoader ğŸ”’

Statics

pgtable ğŸ”’ âš

Functions

get_kernel_elf ğŸ”’

init_percpu ğŸ”’

map_page_range ğŸ”’

Map the specified virtual memory region at the given physical address. This will fail if the caller specifies a virtual address region that is already mapped.

panic ğŸ”’

prepare_kernel_image ğŸ”’

setup_env ğŸ”’ âš

setup_stage2_allocator ğŸ”’

shutdown_percpu ğŸ”’ âš

Release all resources in the PerCpu instance associated with the current CPU.

stage2_main

switch_to_kernel ğŸ”’ âš