Crate stage2

Modules

boot_stage2

Statics

ap_flag ğŸ”’ âš

pgtable ğŸ”’ âš

Functions

check_launch_info ğŸ”’

get_svsm_config ğŸ”’

init_percpu ğŸ”’

load_elf_segment ğŸ”’

Loads a single ELF segment and returns its virtual memory region.

load_igvm_params ğŸ”’

Loads the IGVM params at the next contiguous location from the loaded kernel image. Returns the virtual and physical memory regions hosting the loaded data.

load_kernel_elf ğŸ”’

Loads the kernel ELF and returns the virtual memory region where it resides, as well as its entry point. Updates the used physical memory region accordingly.

map_and_validate ğŸ”’

Map and validate the specified virtual memory region at the given physical address.

panic ğŸ”’

prepare_heap ğŸ”’

Maps any remaining memory between the end of the kernel image and the end of the allocated kernel memory region as heap space. Exclude any memory reserved by the configuration.

setup_env ğŸ”’

setup_stage2_allocator ğŸ”’

shutdown_percpu ğŸ”’ âš

Release all resources in the PerCpu instance associated with the current CPU.

stage2_main