Crate stage2

Modules

boot_stage2

Structs

KernelHeap

Statics

pgtable ğŸ”’ âš

Functions

copy_page_to_kernel ğŸ”’ âš

Safety

init_percpu ğŸ”’

load_elf_segment ğŸ”’ âš

Loads a single ELF segment and returns its virtual memory region.

load_igvm_params ğŸ”’

Loads the IGVM params. Returns the virtual and physical memory regions containing the loaded data.

load_kernel_elf ğŸ”’

Loads the kernel ELF and returns the virtual memory region where it resides, as well as its entry point. Updates the used physical memory region accordingly.

map_and_validate ğŸ”’ âš

Map and validate the specified virtual memory region at the given physical address.

panic ğŸ”’

prepare_heap ğŸ”’

Maps any remaining memory between the end of the kernel image and the end of the allocated kernel memory region as heap space. Exclude any memory reserved by the configuration.

setup_env ğŸ”’ âš

setup_stage2_allocator ğŸ”’

shutdown_percpu ğŸ”’ âš

Release all resources in the PerCpu instance associated with the current CPU.

stage2_main